Consolidate all example tests with their primary tests.
