+++ test/Elements.Tests/ColumnTest.cs
@@ -6,14 +6,19 @@ namespace Elements.Tests
 {
     public class ColumnTest : ModelTest
     {        
-        [Fact]
-        public void Column()
+        [Fact, Trait("Category", "Examples")]
+        public void Example()
         {
             this.Name = "Elements_Column";
 
+            // <example>
+            // Create a framing type.
             var profile = WideFlangeProfileServer.Instance.GetProfileByType(WideFlangeProfileType.W10x100);
+
+            // Create a column.
             var column = new Column(Vector3.Origin, 3.0, profile, BuiltInMaterials.Steel);
-            
+            // </example>
+
             this.Model.AddElement(column);
         }
     }
