+++ test/Elements.Tests/ElementInstanceTests.cs
@@ -6,7 +6,7 @@ namespace Elements.Tests
 {
     public class ElementInstanceTests : ModelTest
     {
-        [Fact]
+        [Fact, Trait("Category", "Examples")]
         public void Instance()
         {
             this.Name = "Elements_ElementInstance";
@@ -16,7 +16,10 @@ namespace Elements.Tests
             var profile = new Profile(Polygon.Rectangle(1.0, 1.0));
             var material = new Material("yellow", Colors.Yellow);
             var line = new Line(Vector3.Origin, new Vector3(5, 5, 5));
-            var testUserElement = new TestUserElement(line, profile, material, true);
+
+            // Set the isElementDefinition parameter to true,
+            // so the element is considered an instance definition.
+            var testUserElement = new TestUserElement(line, profile, material, isElementDefinition: true);
 
             // Create instances of that element
             // with varied transforms.
