+++ test/Elements.Tests/TopographyTests.cs
@@ -22,6 +22,24 @@ namespace Elements.Tests
             this._output = output;
         }
 
+        [Fact, Trait("Category", "Examples")]
+        public void Example()
+        {
+            this.Name = "Elements_Topography";
+            
+            // <example>
+            // Read topo elevations from a file.
+            var data = JsonConvert.DeserializeObject<Dictionary<string,double[]>>(File.ReadAllText("./elevations.json"));
+            var elevations = data["points"];
+            var tileSize = WebMercatorProjection.GetTileSizeMeters(15);
+
+            // Create a topography.
+            var topo = new Topography(Vector3.Origin, tileSize, elevations);
+            // </example>
+
+            this.Model.AddElement(topo);
+        }
+
         [Fact]
         public void Simple()
         {
